### 2020-06-08 #2309 @arfon
@morganjwilliams - will this submission eventually be part of the pyOpenSci GitHub organization? If so, it might be worth waiting until that last step is complete before processing this paper so that the URLs to the software are correct?
