### 2020-06-09 #2309 @morganjwilliams
@arfon migration to the pyOpenSci org isn't planned at this stage, so the links should be OK.
