### 2020-06-09 #2314 @whedon
[REVIEW]: pyrolite: Python for geochemistry
**Submitting author:** @morganjwilliams (<a href="http://orcid.org/0000-0003-4764-9555">Morgan Williams</a>)
**Repository:** <a href="https://github.com/morganjwilliams/pyrolite" target ="_blank">https://github.com/morganjwilliams/pyrolite</a>
**Version:** 0.2.7
