### 2020-06-09 #2314 @arfon
Accepting as part of pyOpenSci (https://github.com/pyOpenSci/software-review/issues/20)
