### 2020-06-09 #2314 @morganjwilliams
Yes, the Zenodo DOI is already active. For this version see [10.5281/zenodo.3877690](https://doi.org/10.5281/zenodo.3877690).
