### 2020-06-09 #2314 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
